requirements.txt
easyocr
streamlit
langchain
langgraph
python-dotenv
pandas
requests
openai
reportlab
google-generativeai
